효율적인 코드로 변경
임시 변수 없이 swap, 2중 반복문 대신 for문 하나 사용
