+++ [정렬] 3월 4일/B10804.py
@@ -1,0 +1,24 @@
+import sys
+input = sys.stdin.readline
+
+def solution():
+    card = [i for i in range(1, 21)]
+
+    for _ in range(10):
+        a, b = map(int, input().split())
+        
+        # for i in range(b, a, -1):
+        #     for j in range(a, i):
+        #         tmp = card[j]
+        #         card[j] = card[j-1]
+        #         card[j-1] = tmp
+        
+        for i in range((b-a)//2 + 1):
+            card[a-1+i], card[b-1-i] = card[b-1-i], card[a-1+i]
+                
+    for c in card:
+        print(c, end=' ')
+    print()
+        
+if __name__ == "__main__":
+    solution()
